+++ NEWS.md
@@ -1,3 +1,3 @@
 # riskscore 0.0.0.9000
 
-* set up pkg skeleton with initial riskscore placeholder
+* set up pkg skeleton with initial `riskscore` placeholder